sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rooch"
version = "0.1.0"
description = "Python SDK for Rooch blockchain"
readme = "README.md"
authors = [
    {name = "Rooch Team", email = "info@rooch.network"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "websockets>=10.4",
    "pycryptodome>=3.15.0",
    "base58>=2.1.1",
    "typing-extensions>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.19.0",
    "black>=22.3.0",
    "isort>=5.10.0",
    "mypy>=0.950",
    "flake8>=4.0.1",
]

[tool.setuptools]
packages = ["rooch"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
